sandbox/python/requirements.txt
optima-sdk @ file://../../../packages/sdk-python
python-dotenv>=1.0.0
